// src/components/10-GitHubProfile/Profile/GithubProfileDetails.tsx
import React from 'react';
import { IProfile } from '../Models/IProfile';
interface IProps{
    profile : IProfile
};
interface IState{};
class GithubProfileDetails extends React.Component<IProps,IState>{
    /**
     *
     */
    constructor(props:IProps) {
        super(props);
        
    }
    render(){
        let {profile} = this.props;
        return(
            <React.Fragment>
             <div className="card">
                <div className="card-header">
                    <span className='badge badge-success mx-2'>{profile.followers}Followers</span>
                    <span className='badge badge-warning mx-2'>{profile.public_repos} Repos</span>
                    <span className='badge badge-secondary mx-2'>{profile.public_gists}Gists</span>
                    <span className='badge-danger mx-2'>{profile.following}Following</span>
                    <span></span>
                </div>
                <div className="card-body">
                    <ul className='list-group'>
                        <li className='list-group-item'>
                            NAME : <span className='font-weight-bold'>{profile.name}</span>
                        </li> 
                        <li className='list-group-item'>
                            Location : <span>{profile.location}</span>
                        </li>
                        <li className='list-group-item'>
                            Email : <span>{profile.email}</span>
                        </li>
                        <li className='list-group-item'>
                            Company : <span>{profile.company}</span>
                        </li>
                        <li className='list-group-item'>
                            Blog : <span>{profile.blog}</span>
                        </li>
                    </ul>
                </div>
             </div>
            </React.Fragment>
        )
    }
}
export default GithubProfileDetails;